move gerTimmings & getMapSrc (& its unit tests) from utils to DetailPage
