+++ helpers/utils.tsx
@@ -1,6 +1,6 @@
 import getConfig from 'next/config';
 
-import { Timming, Restaurant } from './types';
+import { Restaurant } from './types';
 
 export const {
   publicRuntimeConfig: {
@@ -49,24 +49,6 @@ export const getRandomListNumbers = (
   return indexArray;
 };
 
-// DETAIL PAGE UTILS
-
-export const getTimmings = (timmingsStr: string) =>
-  timmingsStr.split(',').reduce((acc: Timming[], next: string) => {
-    const firstDayIndex = next.indexOf('(');
-    const lastDayIndex = next.indexOf(')');
-
-    if (next.includes('(') && next.includes(')')) {
-      acc.push({
-        id: next,
-        day: next.slice(firstDayIndex + 1, lastDayIndex),
-        schedule: next.slice(0, firstDayIndex - 1),
-      });
-    }
-
-    return acc;
-  }, []);
-
 //const checkAlphanumeric = (name: string) => !!name.match(/^[0-9a-zA-Z]+$/);
 
 export const getAlphanumericText = (text: string) =>
@@ -83,13 +65,6 @@ export const getFormattedUrlText = (text: string, isPath = false) => {
 
     return `${acc}${concatenator}${formattedText(next)}`;
   }, '');
-};
-
-export const getMapSrc = (name: string, location: string) => {
-  const urlName = getFormattedUrlText(name);
-  const urlLocation = getFormattedUrlText(location);
-
-  return `https://www.google.com/maps/embed/v1/place?key=${process.env.NEXT_PUBLIC_EMBED_KEY}&q=${urlName}-${urlLocation},Dublin&zoom=16`;
 };
 
 // RELATED RESTAURANTS
